pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "EFMCalculator2"
authors = [
    { name = "croots", email = "croots@utexas.edu" },
    { name = "coolbears", email = "sld3379@utexas.edu" },
    { name = "kevin99111", email = "kevinyang260@gmail.com" },
    { name = "avkatre", email = "avyaykatre@gmail.com" },
]
maintainers = [{ name = "croots", email = "croots@utexas.edu" }]
description = "A webapp and command line utility for identifying mutational hotspots"
# version = Managed by setuptools-scm
dynamic = ["version"]
requires-python = ">=3.12"
readme = "README.md"
keywords = [
    "science",
    "biology",
    "evolution",
    "synthetic biology",
    "synbio",
    "bioinformatics",
    "bioengineering",
]
dependencies = [
    "pandas",
    "polars==1.19",
    "progress",
    "biopython",
    "statsmodels",
    "rich",
    "pyarrow",
    "bokeh==2.4.3",
    "numpy<2",
    "streamlit==1.43.0",
    "streamlit-extras",
    "streamlit-aggrid",
    "streamlit_javascript",
]
license = { text = "GPL-3.0" }

classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

[project.urls]
Homepage = "https://github.com/barricklab/efm-calculator2"
Documentation = "https://github.com/barricklab/efmc-alculator2"
Issues = "https://github.com/barricklab/efm-calculator2/issues"

[project.scripts]
efmcalculator2 = "efmcalculator2.cli:main"
efmcalculator2-webapp = "efmcalculator2.webapp.bootstrap_streamlit:bootstrap_streamlit"
efmcalculator = "efmcalculator2.cli:main"
efmcalculator-webapp = "efmcalculator2.webapp.bootstrap_streamlit:bootstrap_streamlit"


[tool.setuptools]
packages = ["efmcalculator2"]

[tool.setuptools_scm]
version_file = "efmcalculator2/_version.py"
version_scheme = "post-release"
local_scheme = "no-local-version"
